apis/v1: Split PrometheusRule types in their own file
